fix: add `static_cast`s to fix build on windows
these openssl apis aren't size_t-compatible
see also https://github.com/cisco/mlspp/blob/c13d725e6cce923f66496ecc3c43d5d3cd831ec2/lib/hpke/src/aead_cipher.cpp#L202-L206

+++ cpp/src/dave/openssl_cryptor.cpp
@@ -67,7 +67,7 @@ bool OpenSSLCryptor::Encrypt(ArrayView<uint8_t> ciphertextBufferOut,
 
     if (additionalData.size() > 0) {
         auto aadResult = EVP_EncryptUpdate(
-          cipherCtx_, nullptr, &ciphertextOutSize, additionalData.data(), additionalData.size());
+          cipherCtx_, nullptr, &ciphertextOutSize, additionalData.data(), static_cast<int>(additionalData.size()));
 
         if (aadResult != 1) {
             DISCORD_LOG(LS_ERROR) << "Failed to update encryption with additional data";
@@ -80,7 +80,7 @@ bool OpenSSLCryptor::Encrypt(ArrayView<uint8_t> ciphertextBufferOut,
                                           ciphertextBufferOut.data(),
                                           &ciphertextOutSize,
                                           plaintextBuffer.data(),
-                                          plaintextBuffer.size());
+                                          static_cast<int>(plaintextBuffer.size()));
 
     if (updateResult != 1) {
         DISCORD_LOG(LS_ERROR) << "Failed to encrypt plaintext";
@@ -130,7 +130,7 @@ bool OpenSSLCryptor::Decrypt(ArrayView<uint8_t> plaintextBufferOut,
 
     if (additionalData.size() > 0) {
         auto aadResult = EVP_DecryptUpdate(
-          cipherCtx_, nullptr, &plaintextOutSize, additionalData.data(), additionalData.size());
+          cipherCtx_, nullptr, &plaintextOutSize, additionalData.data(), static_cast<int>(additionalData.size()));
 
         if (aadResult != 1) {
             DISCORD_LOG(LS_ERROR) << "Failed to update decryption with additional data";
@@ -143,7 +143,7 @@ bool OpenSSLCryptor::Decrypt(ArrayView<uint8_t> plaintextBufferOut,
                                           plaintextBufferOut.data(),
                                           &plaintextOutSize,
                                           ciphertextBuffer.data(),
-                                          ciphertextBuffer.size());
+                                          static_cast<int>(ciphertextBuffer.size()));
     if (updateResult != 1) {
         DISCORD_LOG(LS_ERROR) << "Failed to decrypt ciphertext";
         PrintSSLErrors();
